+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* [PR #367]: Changes the current goal algorithm to increase by a 25% ratio (Issue 365)
 * [PR #366]: Change petition days remaining message (Issue 364)
 
 ## [1.15.0] - 11/05/2017